Codechange: split some files to keep the ParameterInfo tables styled
parameter_info_table is now outside of validation, to keep the manual
spacing. This required moving the classes over multiple files, to
avoid import cycles and keep everything working.
